Create fitbit.py
Hi Carlo,

## TLDR - Maybe made the fitbit component support multiple devices.  (see disclaimer and instructions below!)

~~~~~~~~~~~~~~~~

## Long version...

I've been looking in to sleep monitoring and was looking at getting fitbits for my other-half and myself, but it's only any use spending the money if I can monitor us both with homeassistant.  As such, I did a bit of research and saw the various threads that show the current component is only working with one device.  So, I decided to have a look at amending it.  Problem being, I can't test it because I don't have the devices to test it on, and I'm not buying the devices until I know it works.

## Disclaimer

I am not a python programmer, I'm just learning the platform now, but I get the logic so I've fumbled through!  It appears to be working as far as I can test it without the devices here.  If this breaks anything I'm really really sorry!  Make sure you follow the instructions below for some safeguarding!

## Instructions

1 - Make a backup copy of your current fitbit.conf somewhere away from your homeassistant instance.

2 - Delete your current local copy of fitbit.conf, and add this PR to your custom_components folder

3 - Change your configuration.yaml (or appropriate package/include) to:

```
sensor:
  - platform: fitbit
    filename: carlo.conf

  - platform: fitbit
    filename: wife.conf
```

plus whatever monitored conditions you want and your clock preference if relevant.

4 -Restart homeassistant

5 - You should get a configurator card for the two entities, copy the contents of your old fitbit.conf in to carlo.conf , sort out the details for your wife's in to hers.

6 - Let me know!!!

## If it all goes wrong

Delete this custom component, reinstate your original fitbit.conf, revert your configuration.yaml (or appropriate include/package).

##Disclaimer 2

If it doesn't work, I might not be able to do much more as I am not a python programmer, but let me know because it might be something I can work out.

If it does work I'll get another couple of people to test and then submit it as an update for the integrated component.